refactor(dom): move playerShowsPageVideo into utils/dom/player
The check is not about captions: any feature that reads or drives the player right after a load or an in-page navigation needs to know whether the player is already showing the page's video. Both caption features keep using it from the new module.
